requirements.txt
asgiref==3.8.1
asttokens==2.4.1
cfgv==3.4.0
colorama==0.4.6
distlib==0.3.8
Django==5.0.3
djangorestframework==3.15.1
executing==2.0.1
factory-boy==3.3.0
Faker==24.4.0
filelock==3.13.3
icecream==2.1.3
identify==2.5.35
nodeenv==1.8.0
platformdirs==4.2.0
pre-commit==3.7.0
Pygments==2.17.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
virtualenv==20.25.1
